Extract `printMemberExpression`
